requirements.txt
anyio==3.6.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
coverage==6.5.0
crispy-bootstrap5==0.6
dj-database-url==1.0.0
Django==4.1.1
django-bootstrap-icons==0.8.1
django-bootstrap4==22.2
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-filter==22.1
django-heroku==0.3.1
django-js-asset==2.0.0
django-on-heroku==1.1.2
django-widget-tweaks==1.4.12
fastapi==0.85.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
Pillow==9.3.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pydantic==1.10.2
requests==2.28.1
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
starlette==0.20.4
typing-extensions==4.3.0
tzdata==2022.2
urllib3==1.26.12
uvicorn==0.18.3
whitenoise==6.2.0
